refactor: move Clone module under MySQL/ namespace
Clone.hs depends entirely on PureMyHA.MySQL.* modules and implements
MySQL CLONE plugin operations, so it belongs in the MySQL directory
alongside Connection, Query, GTID, Auth, and TLS.

Closes #147
